+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Allow passing raw JSON string, JSON file path or URL to `Connection.download()`,
+  `Connection.execute()` and `Connection.create_job()`
+
+
 ### Changed
 
 - `DataCube.execute_batch()` now also guesses the output format from the filename,